refactor: 메인 메뉴와 역 관리 메뉴 enum을 별도 패키지에서 따로 관리
