+++ src/main/java/subway/menus/MainMenu.java
@@ -1,11 +1,11 @@
-package subway.views.mainviews;
+package subway.menus;
 
-enum MainMenu {
+public enum MainMenu {
     STATION_MANAGEMENT("1", "역 관리"),
     LINE_MANAGEMENT("2", "노선 관리"),
     SECTION_MANAGEMENT("3", "구간 관리"),
     SHOW_SUBWAYS("4", "지하철 노선도 출력"),
-    EXIT_PROGRAM("Q","종료");
+    EXIT_PROGRAM("Q", "종료");
 
     private String optionalItem;
     private String itemDescription;
